+++ Runtime/FrameResizeAlgorithm.cs
@@ -1,0 +1,8 @@
+namespace WindowCapture
+{
+    public enum FrameResizeAlgorithm
+    {
+        Nearest = 0,
+        Bilinear = 1
+    }
+}